# openapi/openapi.interno.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {},
    "description": "Contrato HTTP generado desde los controladores de la aplicación.",
    "title": "LuxeBorealCRM API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/asistente/categorias/orden": {
      "put": {
        "operationId": "ordenarCategoriasCaso",
        "parameters": [
          {
            "description": "Anti-CSRF: siempre `1`.",
            "in": "header",
            "name": "X-Luxe-Csrf",
            "required": true,
            "schema": {
              "enum": [
                "1"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ids": {
                    "items": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "maxItems": 500,
                    "type": "array"
                  }
                },
                "required": [
                  "ids"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "categorias": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                            "type": "string"
                          },
                          "nombre": {
                            "type": "string"
                          },
                          "orden": {
                            "maximum": 9007199254740991,
                            "minimum": -9007199254740991,
                            "type": "integer"
                          },
                          "totalCasos": {
                            "maximum": 9007199254740991,
                            "minimum": -9007199254740991,
                            "type": "integer"
                          }
                        },
                        "required": [
                          "id",
                          "nombre",
                          "orden",
                          "totalCasos"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "categorias"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Las categorías en el orden nuevo."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errores": {
                      "items": {
                        "properties": {
                          "campo": {
                            "type": "string"
                          },
                          "problema": {
                            "enum": [
                              "falta",
                              "formato",
                              "valor"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "campo",
                          "problema"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "codigo"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "El cuerpo no cumple el esquema (validacion-fallida)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errores": {
                      "items": {
                        "properties": {
                          "campo": {
                            "type": "string"
                          },
                          "problema": {
                            "enum": [
                              "falta",
                              "formato",
                              "valor"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "campo",
                          "problema"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "codigo"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Sin sesión válida (peticion-no-autenticada)."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errores": {
                      "items": {
                        "properties": {
                          "campo": {
                            "type": "string"
                          },
                          "problema": {
                            "enum": [
                              "falta",
                              "formato",
                              "valor"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "campo",
                          "problema"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "codigo"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Sin el rol admin (rol-insuficiente) o sin el encabezado anti-CSRF (encabezado-csrf-ausente)."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errores": {
                      "items": {
                        "properties": {
                          "campo": {
                            "type": "string"
                          },
                          "problema": {
                            "enum": [
                              "falta",
                              "formato",
                              "valor"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "campo",
                          "problema"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "codigo"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "La lista no coincide con las categorías existentes (orden-categorias-invalido)."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Reordena las categorías con la lista completa de sus identificadores",
        "tags": [
          "Asistente"
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "luxe_sesion",
        "type": "apiKey"
      }
    }
  }
}